Pin WebOb version to < 1.4 due to removed 'exception' property
In version 1.4 of WebOb the HTTPException.exception property was
removed. This property is used by Pylons. Since we don't expect a new
Pylons release the only solution is to pin the WebOb version.

From WebOb source code:
As of WebOb 1.2, raise the HTTPException instance directly
instead of raising the result of 'HTTPException.exception'

Change-Id: I2104dcf56a3a8f79d5158e642417d14884b9d00c