Refactor in order to use linotpselfservice.lib.network instead of httplib
httplib caused a number of errors which were probably related to HTTP
Keep-Alive and the remote server closing the connection, which was not
handled well locally.

Also httplib does not provide options to verify the SSL/TLS certificate
of the remote server, which is a requirement for us.

linotpselfservice.lib.network is a wrapper for the requests library.

Config changes
--------------

The parameters of the selfservice.ini and who.ini files were adapted so
they use the same names in both, and an additional parameter for a
server certificate was added.

    linotp_url = https://LINOTP.SERVER.COM/
    client_key = %(here)s/client_key.key
    client_cert = %(here)s/client_cert.pem
    server_cert = %(here)s/server_cert.pem

Change-Id: I9ea849bb65742d261b0fe15071b215ddb3adbae2